Avoid problems with explicit_bzero() and LibreSSL
sigbak's explicit_bzero() compat implementation calls OPENSSL_cleanse(),
but LibreSSL's OPENSSL_cleanse() in turn calls explicit_bzero(). This
results in an infinite recursion.

Avoid this by not providing the explicit_bzero() compat implementation
if sigbak is linked against LibreSSL. This will cause sigbak to use the
explicit_bzero() implementation from LibreSSL (which is visible) or from
libc (if it has one).

This fixes a segfault on Void Linux with musl if HAVE_EXPLICIT_BZERO is
not defined.

The fix is from openssh-portable via opensmtpd-portable:

https://github.com/OpenSMTPD/OpenSMTPD/commit/f948b923873a93472dea9b786cf60a3472b0ddc8