Don't error out on missing attachments
It is possible for an attachment to be missing from a backup even when
its transfer state indicates it should be present.

It seems this can happen when an attachment in Signal's app storage on
the phone becomes unreadable so Signal can't add it to the backup.

Instead of erroring out, print a warning on stderr.

Reported by @kadogo in pull request #3. Thanks!